+++ Google/GoogleAnalyticsQueryService.php
@@ -16,20 +16,27 @@ use Piwik\DataTable\Row;
 use Piwik\Date;
 use Piwik\Db;
 use Piwik\Metrics;
+use Piwik\Piwik;
 use Piwik\Site;
 use Piwik\Tracker\GoalManager;
 use Psr\Log\LoggerInterface;
 
 class GoogleAnalyticsQueryService
 {
-    const MAX_ATTEMPTS = 30;
+    const DEFAULT_MAX_ATTEMPTS = 30;
     const MAX_BACKOFF_TIME = 60;
     const PING_MYSQL_EVERY = 25;
+    const DEFAULT_MIN_BACKOFF_TIME = 2; // start at 2s since GA seems to have trouble w/ the 10 requests per 100s limit w/ 1
 
     private static $problematicMetrics = [
         'ga:users',
         'ga:hits',
     ];
+
+    /**
+     * @var int
+     */
+    private $maxAttempts = self::DEFAULT_MAX_ATTEMPTS;
 
     /**
      * @var \Google_Service_Analytics
@@ -54,7 +61,7 @@ class GoogleAnalyticsQueryService
     /**
      * @var int
      */
-    private $currentBackoffTime = 1;
+    private $currentBackoffTime = self::DEFAULT_MIN_BACKOFF_TIME;
 
     private $pingMysqlEverySecs;
 
@@ -146,10 +153,11 @@ class GoogleAnalyticsQueryService
 
         $request = $this->googleQueryObjectFactory->make($this->viewId, $date, $metricNamesChunk, $options);
 
-        $this->currentBackoffTime = 1;
+        $lastGaError = null;
+        $this->currentBackoffTime = self::DEFAULT_MIN_BACKOFF_TIME;
 
         $attempts = 0;
-        while ($attempts < self::MAX_ATTEMPTS) {
+        while ($attempts < $this->maxAttempts) {
             try {
                 $this->issuePointlessMysqlQuery();
 
@@ -159,7 +167,8 @@ class GoogleAnalyticsQueryService
 
                 if (empty($result)) {
                     ++$attempts;
-                    sleep(1);
+
+                    $this->backOff();
 
                     $this->logger->info("Google Analytics API returned null for some reason, trying again...");
 
@@ -172,6 +181,11 @@ class GoogleAnalyticsQueryService
                     'message' => $ex->getMessage(),
                 ]);
 
+                /**
+                 * @ignore
+                 */
+                Piwik::postEvent('GoogleAnalyticsImporter.onApiError', [$ex]);
+
                 if ($ex->getCode() == 403 || $ex->getCode() == 429) {
                     if (strpos($ex->getMessage(), 'daily') !== false) {
                         throw new DailyRateLimitReached();
@@ -180,20 +194,36 @@ class GoogleAnalyticsQueryService
                     ++$attempts;
 
                     $this->logger->debug("Waiting {$this->currentBackoffTime}s before trying again...");
-                    $this->sleep($this->currentBackoffTime);
 
-                    $this->currentBackoffTime = min(self::MAX_BACKOFF_TIME, $this->currentBackoffTime * 2);
+                    $this->backOff();
+                } else if ($this->isIgnorableException($ex)) {
+                    ++$attempts;
+
+                    $this->logger->info("Google Analytics API returned an ignorable or temporary error: {$ex->getMessage()}. Waiting {$this->currentBackoffTime}s before trying again...");
+
+                    $this->backOff();
                 } else if ($ex->getCode() >= 500) {
                     ++$attempts;
-                    $this->logger->info("Google Analytics API returned error: {$ex->getMessage()}. Waiting one minute before trying again...");
-                    $this->sleep(60);
+
+                    $this->logger->info("Google Analytics API returned error: {$ex->getMessage()}. Waiting {$this->currentBackoffTime}s before trying again...");
+
+                    $messageContent = @json_decode($ex->getMessage(), true);
+                    if (isset($messageContent['error']['message'])) {
+                        $lastGaError = $messageContent['error']['message'];
+                    }
+
+                    $this->backOff();
                 } else {
                     throw $ex;
                 }
             }
         }
 
-        throw new \Exception("Failed to reach GA after " . self::MAX_ATTEMPTS . " attempts. Restart the import later.");
+        $message = "Failed to reach GA after " . $this->maxAttempts . " attempts. Restart the import later.";
+        if (!empty($lastGaError)) {
+            $message .= ' Last GA error message: ' . $lastGaError;
+        }
+        throw new \Exception($message);
     }
 
     /**
@@ -202,6 +232,14 @@ class GoogleAnalyticsQueryService
     public function setOnQueryMade($onQueryMade)
     {
         $this->onQueryMade = $onQueryMade;
+    }
+
+    /**
+     * @param int $maxAttempts
+     */
+    public function setMaxAttempts($maxAttempts)
+    {
+        $this->maxAttempts = $maxAttempts;
     }
 
     /**
@@ -223,5 +261,29 @@ class GoogleAnalyticsQueryService
 
             $this->issuePointlessMysqlQuery();
         }
+    }
+
+    private function backOff()
+    {
+        $this->sleep($this->currentBackoffTime);
+        $this->currentBackoffTime = min(self::MAX_BACKOFF_TIME, $this->currentBackoffTime * 2);
+    }
+
+    private function isIgnorableException(\Exception $ex)
+    {
+        if ($ex->getCode() !== 400) {
+            return false;
+        }
+
+        $messageContent = @json_decode($ex->getMessage(), true);
+        if (empty($messageContent['error']['message'])) {
+            return false;
+        }
+
+        if (strpos($messageContent['error']['message'], 'Unknown metric') === 0) {
+            return true;
+        }
+
+        return false;
     }
 }