+++ Google/GoogleCustomDimensionMapper.php
@@ -8,9 +8,9 @@
 
 namespace Piwik\Plugins\GoogleAnalyticsImporter\Google;
 
-
 use Piwik\Plugins\CustomDimensions\Dimension\Name;
 use Piwik\Plugins\GoogleAnalyticsImporter\CannotImportCustomDimensionException;
+use Piwik\Plugins\GoogleAnalyticsImporter\OutOfCustomDimensionsException;
 
 class GoogleCustomDimensionMapper
 {
@@ -22,7 +22,7 @@ class GoogleCustomDimensionMapper
             'extractions' => [],
             'case_sensitive' => true,
             'scope' => $this->mapScope($gaCustomDimension),
-            'active' => $gaCustomDimension->getActive(),
+            'active' => (bool) $gaCustomDimension->getActive(),
         ];
 
         $blockedChars = Name::getBlockedCharacters();
@@ -31,7 +31,7 @@ class GoogleCustomDimensionMapper
         return $result;
     }
 
-    private function mapScope(\Google_Service_Analytics_CustomDimension $gaCustomDimension)
+    public function mapScope(\Google_Service_Analytics_CustomDimension $gaCustomDimension)
     {
         $scope = $gaCustomDimension->getScope();
         switch (strtolower($scope)) {
@@ -41,6 +41,38 @@ class GoogleCustomDimensionMapper
                 return 'visit';
             default:
                 throw new CannotImportCustomDimensionException($gaCustomDimension, 'unsupported scope, "' . $scope . '"');
+        }
+    }
+
+
+    public function checkCustomDimensionCount($availableScopes, $gaCustomDimensions, $extraDimensions)
+    {
+        foreach ($availableScopes as $scope) {
+            $requestedScopes = 0;
+            foreach ($extraDimensions as $extraDimension) {
+                if ($extraDimension['dimensionScope'] == $scope['value']) {
+                    ++$requestedScopes;
+                }
+            }
+
+            /** @var \Google_Service_Analytics_CustomDimension $gaCustomDimension */
+            foreach ($gaCustomDimensions->getItems() as $gaCustomDimension) {
+                try {
+                    $mappedScope = $this->mapScope($gaCustomDimension);
+                } catch (CannotImportCustomDimensionException $ex) {
+                    continue;
+                }
+
+                if ($mappedScope == $scope['value']) {
+                    ++$requestedScopes;
+                }
+            }
+
+            $availableScopes = (int) $scope['numSlotsAvailable'];
+
+            if ($requestedScopes > $availableScopes) {
+                throw new OutOfCustomDimensionsException($requestedScopes, $availableScopes, $scope['value']);
+            }
         }
     }
 }