+++ angularjs/import-scheduler/import-scheduler.controller.js
@@ -20,6 +20,7 @@
         vm.isStartingImport = false;
         vm.extraCustomDimensions = [];
         vm.isVerboseLoggingEnabled = false;
+        vm.ignoreCustomDimensionSlotCheck = false;
         vm.startImport = startImport;
 
         function startImport() {
@@ -36,6 +37,7 @@
 
             vm.isStartingImport = true;
 
+            var forceCustomDimensionSlotCheck = !vm.ignoreCustomDimensionSlotCheck;
             piwikApi.withTokenInUrl();
             return piwikApi.post({
                 module: 'GoogleAnalyticsImporter',
@@ -49,7 +51,8 @@
                 isMobileApp: vm.isMobileApp ? '1' : '0',
                 timezone: vm.timezone,
                 extraCustomDimensions: vm.extraCustomDimensions,
-                isVerboseLoggingEnabled: vm.isVerboseLoggingEnabled ? '1' : '0'
+                isVerboseLoggingEnabled: vm.isVerboseLoggingEnabled ? '1' : '0',
+                forceCustomDimensionSlotCheck: forceCustomDimensionSlotCheck ? '1' : '0'
             }, { token_auth: piwik.token_auth })['finally'](function () {
                 window.location.reload();
             });