+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 ## Changelog
 
+# 1.5.4
+
+Changes:
+* Abort on all errors and report when it is due to a insufficient privileges exception.
+* Ignore unknown metric exceptions.
+* Better debug exception messages when errors occur during controller actions.
+
+Bug Fixes:
+* Handle GA API active custom dimension value of empty string properly.
+* Handle invalid max end date configuration.
+* Check for custom dimension slots before importing & allow ignoring extra custom dimensions
+
+# 1.5.3
+
+Changes:
+* Fix referrers table subtable in imported reports so link is correct (only affects newly imported reports).
+* Show last GA error if there was one when cannot reach GA API fails repeatedly.
+* Use exponential backoff for when GA API backend fails.
+* Fix forum link in error message.
+
 # 1.5.2
 
 Changes: